pyproject.toml
[project]
name = "grist-custom-forms"
version = "0.1.0"
description = "API server for custom Grist forms"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "gunicorn>=22.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.scripts]
serve = "app:app.run"
